+++ ide/js/pow.ts
@@ -81,11 +81,12 @@ const pow = (function(){
     }
 
     /**
-     * Return a built pipeline definition (pipeline.json)
+     * Return a built pipeline definition
      * @param {string} path Path to the pipeline ("!pipeline1" for default workspace)
      * @returns {Promise} Promise with a POWResult
      */
     async function pipeline(path) {
+        // If we don't have a !* or a */* path, add the ! prefix
         if (!path.match(/^!/) && !path.match(/[\\/]/)) path = "!"+path;
         return execStrictJSON(`pow pipeline "${path}" export`);
     }
@@ -150,7 +151,7 @@ const pow = (function(){
      * @returns {Promise} Promise with a POWResult containing an object (pipeline result)
      */
     async function run(path) {
-        return execStrictJSON(`pow run "${path}"`);
+        return execStrictJSON(`pow run "${path}" $null export`);
     }
 
     /**