+++ tests/pipeline-manager.tests.js
@@ -1,6 +1,6 @@
 let pipelineManager = require("../ide/js/pipeline-manager.js");
 (function pipelineManagerTest(console, args) {
-    
+
     let verbose = args.indexOf("verbose")>=0?true:false;
     let FUNC = require("./functions")(verbose);
     let assert = FUNC.assert;
@@ -25,12 +25,9 @@ let pipelineManager = require("../ide/js/pipeline-manager.js");
     const path = require("path");
     // @ts-ignore
     let testPipeline = fs.readFileSync(path.resolve(__dirname, "../examples/pipeline1/pipeline.json"), "utf8").trim();
-    //testPipeline = { "id": "pipeline1", "name": "Send mail to young voters", "description": "Read voters.txt, get all voters under 21yrs of age and send them an email", "parameters": { "DataSource": { "default": ".\\data\\voters.txt", "type": "string" }, "p2": { "default": "{Get-Date}" } }, "globals": { "foo": "bar" }, "checks": { "run": "some checks that we have all we need (e.g. ./data/voters.txt) to run?" }, "input": {}, "output": {}, "steps": [ { "id":"A1", "name":"Read Voters File", "reference":"../components/ReadFile", "input":"", "parameters": { "Path": "{$PipelineParams.DataSource}" } }, { "id":"B1", "name":"Convert2JSON", "reference":"../components/CSV2JSON", "input": "A", "parameters": { "Delimiter": "|", "Header": "{\"name\", \"age\", \"email\", \"source\"}" } }, { "id":"C1", "name":"Select Name and Email", "reference":"../components/SelectFields", "input": "B", "parameters": { "Fields": "{\"name\", \"age\", \"email\"}" } } ] };
     // @ts-ignore
     let testComponent = JSON.parse(fs.readFileSync(path.resolve(__dirname, "./CSV2JSON.json"), "utf8").trim());
     let testComponent2 = JSON.parse(fs.readFileSync(path.resolve(__dirname, "./FileList.json"), "utf8").trim());
-    //console.log(testComponent);process.exit();
-    //testComponent = { "name": "Convert CSV data to JSON format", "description": "Accepts tabular CSV data and return contents as a JSON Array", "parameters": { "FieldSeparator": { "type": "string", "default": ",", "description": "Specifies the field separator. Default is a comma." } }, "input":"text/csv", "output":"json/array"};
 
     try{
 
@@ -39,7 +36,7 @@ let pipelineManager = require("../ide/js/pipeline-manager.js");
         assert(PM.columnCount()===COLS.length, "We have the right number of columns ("+COLS.length+")")
         assert(PM.rowCount()===9, "We have the right number of rows (9)")
         assert(PM.nextRow("A")===1, "Next empty row in column A is 1")
-        
+
         // Getting and Adding Steps
         assert(()=>PM.getStep(1, 1).reference === null, "Step A1 is initialized")
         assert(()=>PM.addComponent("1", 1, testComponent).reference.match(/csv2json/), "Step A1 is set to CSV2JSON")
@@ -60,7 +57,7 @@ let pipelineManager = require("../ide/js/pipeline-manager.js");
         assert(()=>PM.moveStep("C3", "A1"), "Step can't be moved over existing step", true);
         assert(()=>PM.getStep("D4").reference===null, "Step D4 should be empty");
         assert(()=>PM.getStep("A1").reference!==null, "Step A1 should not be empty");
-        assert(()=>PM.getStep("C3").reference!==null, "Step C3 should not be empty"); 
+        assert(()=>PM.getStep("C3").reference!==null, "Step C3 should not be empty");
         PM.moveStep("C3", "D4")
         let c3 = PM.getStep("C3");
         let d4 = PM.getStep("D4");
@@ -76,24 +73,19 @@ let pipelineManager = require("../ide/js/pipeline-manager.js");
         assert(()=>PM.getStep("B1").reference.match(/csv2json/i), "Test pipeline step B1 is CSV2JSON")
         assert(()=>PM.getStep("C1").reference.match(/selectfields/i), "Test pipeline step C1 is SelectFields")
         assert(()=>typeof PM.getStep("A1").parameters.Path === "string", "Step A1 of imported pipeline has a 'Path' parameter")
-        
+
         // Export
         let myPipeline = PM.export();
         assert(()=>myPipeline.steps.length===3, "Pipeline export has 3 steps")
         let pipelineDef = PM.getDefinition();
         assert(()=>myPipeline.id===pipelineDef.id, "Export has same id")
-        //let pipelineStr = JSON.stringify(myPipeline);
-        //fs.writeFileSync("C:\\temp\\pipeline.json", pipelineStr, "utf-8");
-
-        //console.log(PM.getStep("C1"))
-        //columnsToString();
 
     } catch(e) {
         FUNC.addFails();
         console.log("TEST cancelled:")
         console.log(e);
     }
-    
+
     if (FUNC.fails() > 0) {
         columnsToString();
         console.log("\x1b[31m", `FAIL: pipeline-manager.js failed ${FUNC.fails()} of ${FUNC.tests()} tests`)